be/requirements.txt
motor
httpx
uvicorn
fastapi
pymongo
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
